# packages/eval-hub/api/openapi/eval-hub.yaml
openapi: 3.0.3
info:
  title: 'EvalHub REST API'
  version: '1.0.0'
  description: 'REST API for EvalHub Backend for Frontend (BFF) using OpenAI SDK v2'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
servers:
  - url: 'http://localhost:4000'
    description: 'Local development server'
security:
  - Bearer: []
paths:
...
  /eval-hub/api/v1/evaluations/jobs/{id}/logs:
    summary: Get evaluation job logs
    description: >-
      Retrieves execution logs for an evaluation job as plain text from the
      EvalHub Server. Supports server-side truncation via tail_lines and
      time-based filtering via since_seconds.
    get:
      tags:
        - Evaluations
      security:
        - Bearer: []
      parameters:
        - $ref: '#/components/parameters/evaluationJobId'
        - $ref: '#/components/parameters/namespace'
        - $ref: '#/components/parameters/tailLines'
        - $ref: '#/components/parameters/timestamps'
        - $ref: '#/components/parameters/sinceSeconds'
      responses:
        '200':
          description: Plain text execution logs
          headers:
            X-Log-Truncated:
              description: Whether server-side byte or timeout limits truncated the log response
              required: true
              schema:
                type: boolean
          content:
            text/plain:
              schema:
                type: string
              example: |
                === Logs for job eval-job-001 ===
                [2026-03-01T09:00:00Z] Starting evaluation...
                [2026-03-01T09:05:00Z] Benchmark truthfulqa_mc1 completed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      operationId: getEvaluationJobLogs
      summary: Get Evaluation Job Logs
      description: >-
        Retrieves execution logs for an evaluation job. Returns plain text
        with optional per-benchmark section headers.
  /eval-hub/api/v1/evaluations/jobs/{id}/benchmarks/{benchmark_index}/logs:
    summary: Get evaluation job benchmark logs
    description: >-
      Retrieves execution logs for a specific benchmark within an evaluation
      job as plain text from the EvalHub Server.
    get:
      tags:
        - Evaluations
      security:
        - Bearer: []
      parameters:
        - $ref: '#/components/parameters/evaluationJobId'
        - $ref: '#/components/parameters/benchmarkIndex'
        - $ref: '#/components/parameters/namespace'
        - $ref: '#/components/parameters/tailLines'
        - $ref: '#/components/parameters/timestamps'
        - $ref: '#/components/parameters/sinceSeconds'
      responses:
        '200':
          description: Plain text execution logs for the specified benchmark
          headers:
            X-Log-Truncated:
              description: Whether server-side byte or timeout limits truncated the log response
              required: true
              schema:
                type: boolean
          content:
            text/plain:
              schema:
                type: string
              example: |
                === Benchmark 0 Logs ===
                [2026-03-01T09:01:00Z] Running benchmark...
                [2026-03-01T09:05:00Z] Benchmark completed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      operationId: getEvaluationJobBenchmarkLogs
      summary: Get Evaluation Job Benchmark Logs
      description: >-
        Retrieves execution logs for a specific benchmark within an
        evaluation job. Returns plain text log output.

  # =============================================================================
  # INTER-BFF: MODEL CATALOG SECURITY ARTIFACTS
  # =============================================================================
...
components:
...
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: '400'
          description: HTTP status code
        message:
          type: string
          example: 'Bad request'
          description: Human-readable error message
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'

  # =============================================================================
  # RESPONSES
  # =============================================================================
  responses:
...
    BadRequest:
      description: Bad Request - Invalid parameters or missing required fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: '400'
              message: 'invalid request parameters'
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Not Found - Requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalServerError:
      description: Internal Server Error - Server encountered an unexpected condition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'

  # =============================================================================
  # PARAMETERS
  # =============================================================================
  parameters:
...
    namespace:
      in: query
      name: namespace
      schema:
        type: string
      required: false
      description: Kubernetes namespace to scope the request (optional)
...
    evaluationJobId:
      in: path
      name: id
      schema:
        type: string
      required: true
      description: Unique identifier for the evaluation job
      example: 'eval-job-001'
    benchmarkIndex:
      in: path
      name: benchmark_index
      schema:
        type: integer
        minimum: 0
      required: true
      description: Zero-based index of the benchmark within the evaluation job
      example: 0
    tailLines:
      in: query
      name: tail_lines
      schema:
        type: integer
        minimum: -1
      required: false
      description: Number of lines from the end of the log to return; -1 requests all available lines subject to server limits
      example: -1
    timestamps:
      in: query
      name: timestamps
      schema:
        type: boolean
      required: false
      description: Whether to include timestamps in log lines
    sinceSeconds:
      in: query
      name: since_seconds
      schema:
        type: integer
        minimum: 0
      required: false
      description: Only return log lines newer than this many seconds
      example: 3600
...
  securitySchemes:
    Bearer:
      type: apiKey
      in: header
      name: x-forwarded-access-token
      description: User access token (use `oc whoami -t` to get your cluster token)
...